Feat: TouchAreaProvider 생성 및 JSDoc으로 주석 추가
