+++ src/App.jsx
@@ -1,4 +1,5 @@
 import { BrowserRouter, Route, Routes } from "react-router-dom";
+import { TouchAreaProvider } from "context/TouchAreaProvider";
 import Landing from "pages/Landing";
 import List from "pages/List";
 import MyPage from "pages/MyPage";
@@ -7,12 +8,14 @@ import "styles/reset.scss";
 
 export default function App() {
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Landing />} />
-        <Route path="/list" element={<List />} />
-        <Route path="/mypage" element={<MyPage />} />
-      </Routes>
-    </BrowserRouter>
+    <TouchAreaProvider>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<Landing />} />
+          <Route path="/list" element={<List />} />
+          <Route path="/mypage" element={<MyPage />} />
+        </Routes>
+      </BrowserRouter>
+    </TouchAreaProvider>
   );
 }
